% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{add_new_col_to_data_table}
\alias{add_new_col_to_data_table}
\title{add new col to data table}
\usage{
add_new_col_to_data_table(original_data_table, new_col, new_col_name)
}
\arguments{
\item{original_data_table}{the original data table}

\item{new_col}{the new col}

\item{new_col_name}{the new col name}
}
\description{
add new col to data table
}
