% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-panel.R
\name{annotate_text_JuJu}
\alias{annotate_text_JuJu}
\title{annotate text JuJu}
\usage{
annotate_text_JuJu(
  label,
  angle = 0,
  size = 12,
  family = "Arial",
  face = "plain",
  panel.background = element_blank(),
  plot.background = element_blank(),
  vjust = 0.5,
  hjust = 0.5,
  plot.margin = margin(0, 0, 0, 0),
  ratio = NULL
)
}
\arguments{
\item{label}{the label}

\item{angle}{the label angle}

\item{size}{the label size}

\item{family}{the label family}

\item{face}{the label face}
}
\description{
annotate text JuJu
}
