% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-panel.R
\name{combine_figures_JuJu}
\alias{combine_figures_JuJu}
\title{combine figures JuJu}
\usage{
combine_figures_JuJu(
  figure_ls,
  col_na,
  row_na,
  label.size = 12,
  label.face = "plain",
  plot.margin = margin(t = 5, b = 5, r = 5, l = 5, unit = "pt"),
  plot.background = element_blank()
)
}
\arguments{
\item{figure_ls}{the figure list}

\item{col_na}{the column names}

\item{row_na}{the row names}
}
\description{
combine figures JuJu
}
