% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ANOVA.R
\name{conduct_ANOVE_JuJu}
\alias{conduct_ANOVE_JuJu}
\title{conduct ANOVE test}
\usage{
conduct_ANOVE_JuJu(dataset_dt, group_col, test_col, normality = TRUE)
}
\arguments{
\item{dataset_dt}{the dataset table}

\item{group_col}{the group col}

\item{test_col}{the testing col}
}
\description{
conduct ANOVE test
}
