% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utills-description_statistical_analysis.R
\name{conduct_description_statistical_analysis_JuJu}
\alias{conduct_description_statistical_analysis_JuJu}
\title{conduct description statistical analysis}
\usage{
conduct_description_statistical_analysis_JuJu(dataset_dt, group_col, test_col)
}
\arguments{
\item{dataset_dt}{the dataset tabel}

\item{group_col}{the group col}

\item{test_col}{the testing col}
}
\description{
conduct description statistical analysis
}
