% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shaprio_test.R
\name{conduct_shaprio_test_JuJu}
\alias{conduct_shaprio_test_JuJu}
\title{conduct shaprio test}
\usage{
conduct_shaprio_test_JuJu(dataset_dt, group_col, test_col)
}
\arguments{
\item{dataset_dt}{the dataset tabel}

\item{group_col}{the group col}

\item{test_col}{the testing col}
}
\description{
conduct shaprio test
}
