% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-t_test.R
\name{conduct_t_test_each_JuJu}
\alias{conduct_t_test_each_JuJu}
\title{conduct t test}
\usage{
conduct_t_test_each_JuJu(
  test_dataset,
  group_col,
  test_mode = "two.sided",
  test_cols = character(),
  var.equal = FALSE,
  paired = FALSE,
  conf.level = 0.95,
  normality = TRUE,
  simplify = FALSE
)
}
\arguments{
\item{test_dataset}{the data table}

\item{group_col}{the column of group symbol}

\item{test_mode}{the test mode}
}
\description{
conduct t test
}
