% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-t_test.R
\name{conduct_t_test_JuJu}
\alias{conduct_t_test_JuJu}
\title{conduct t test}
\usage{
conduct_t_test_JuJu(
  dataset_dt,
  group_col,
  group_symbol_1,
  group_symbol_2,
  test_mode = "two.sided",
  test_cols = character()
)
}
\arguments{
\item{dataset_dt}{the data table}

\item{group_col}{the column of group symbol}

\item{group_symbol_1}{the group symbol 1}

\item{group_symbol_2}{the group symbol 2}

\item{test_mode}{the test mode}

\item{test_cols}{the columns to test}
}
\description{
conduct t test
}
