% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{convet_dataset_to_short_form_JuJu}
\alias{convet_dataset_to_short_form_JuJu}
\title{convert a dataset to short form}
\usage{
convet_dataset_to_short_form_JuJu(dt, col, save_cols)
}
\arguments{
\item{dt}{the data.table}

\item{col}{the change col}

\item{save_cols}{the saving cols}
}
\description{
convert a dataset to short form
}
