% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-t_test_barchart.R
\name{.conduct_loop_adding_to_plotting_dataset}
\alias{.conduct_loop_adding_to_plotting_dataset}
\title{conduct loop adding to plotting dataset}
\usage{
.conduct_loop_adding_to_plotting_dataset(
  comp_pairs_ls,
  group_na,
  max_d,
  sub_plotting_dataset_by_group,
  sub_plotting_dataset_by_comparition
)
}
\arguments{
\item{comp_pairs_ls}{compared pairs list}

\item{group_na}{the group name}

\item{max_d}{the max distance}

\item{sub_plotting_dataset_by_group}{the sub plotting dataset unique by group}

\item{sub_plotting_dataset_by_comparition}{the sub plotting dataset unique by comparition}
}
\description{
conduct loop adding to plotting dataset
}
