% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.convert_p.value_to_significant_stars}
\alias{.convert_p.value_to_significant_stars}
\title{convert p.value to significant stars}
\usage{
.convert_p.value_to_significant_stars(p.value, significant_level)
}
\arguments{
\item{p.value}{the p.value}

\item{significant_level}{the significant level of p.value}
}
\description{
convert p.value to significant stars
}
