% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.convert_statistic_result_to_plotting_dataset}
\alias{.convert_statistic_result_to_plotting_dataset}
\title{convert statistic result to plotting dataset}
\usage{
.convert_statistic_result_to_plotting_dataset(
  statistic_result,
  test.type = "t_test"
)
}
\arguments{
\item{statistic_result}{the statistic result in data.table form}

\item{test.type}{the test type}
}
\description{
convert statistic result to plotting dataset
}
