% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-t_test_barchart.R
\name{.generate_additional_dataset_from_plotting_dataset}
\alias{.generate_additional_dataset_from_plotting_dataset}
\title{generate additional dataset from plotting dataset}
\usage{
.generate_additional_dataset_from_plotting_dataset(
  plotting_dataset,
  test.type = "t_test",
  group_order = character()
)
}
\arguments{
\item{plotting_dataset}{the plotting dataset}

\item{test.type}{the test type}

\item{group_order}{the group order}
}
\description{
generate additional dataset from plotting dataset
}
