% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-t_test_barchart.R
\name{.generate_additional_dataset_from_sub_plotting_dataset_for_t_test}
\alias{.generate_additional_dataset_from_sub_plotting_dataset_for_t_test}
\title{generate additional dataset from sub plotting dataset for t test}
\usage{
.generate_additional_dataset_from_sub_plotting_dataset_for_t_test(
  sub_plotting_dataset,
  group_order = character()
)
}
\arguments{
\item{sub_plotting_dataset}{the sub plotting dataset}

\item{group_order}{the group order}
}
\description{
generate additional dataset from sub plotting dataset for t test
}
