% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-t_test_barchart.R
\name{.generate_segment_dataset_from_plotting_dataset}
\alias{.generate_segment_dataset_from_plotting_dataset}
\title{generate segment dataset from plotting dataset}
\usage{
.generate_segment_dataset_from_plotting_dataset(
  plotting_dataset,
  test.type = "t_test"
)
}
\arguments{
\item{plotting_dataset}{the plotting dataset}

\item{test.type}{the test type}
}
\description{
generate segment dataset from plotting dataset
}
