% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-integrated_analysis_pipeline.R
\name{.pipeline_ANOVE}
\alias{.pipeline_ANOVE}
\title{pipeline ANOVE}
\usage{
.pipeline_ANOVE(
  dataset_dt,
  group_col,
  test_col,
  alpha = 0.05,
  normality = TRUE
)
}
\arguments{
\item{dataset_dt}{the dataset tabel}

\item{group_col}{the group col}

\item{test_col}{the testing col}
}
\description{
pipeline ANOVE
}
