% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-t_test_barchart.R
\name{.seperate_compared_pairs}
\alias{.seperate_compared_pairs}
\title{seperate compared pairs}
\usage{
.seperate_compared_pairs(comp_pairs)
}
\arguments{
\item{comp_pairs}{the comparing pairs}
}
\description{
seperate compared pairs
}
