% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-geom_segment_JuJu.R
\name{.transform_data_JuJu}
\alias{.transform_data_JuJu}
\title{transform data JuJu}
\usage{
.transform_data_JuJu(data, geom_symbol)
}
\arguments{
\item{data}{the data}

\item{geom_symbol}{the geom symbol}
}
\description{
transform data JuJu
}
