% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export_data.table.R
\name{export_as_excel_JuJu}
\alias{export_as_excel_JuJu}
\title{export as excel}
\usage{
export_as_excel_JuJu(dt_ls, wb_na, path, new = TRUE)
}
\arguments{
\item{dt_ls}{the data.table list}

\item{wb_na}{the workbook name}
}
\description{
export as excel
}
