% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-GeomSegmentJuJu.R
\name{geom_segment_JuJu}
\alias{geom_segment_JuJu}
\title{geom_segment_JuJu}
\usage{
geom_segment_JuJu(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = position_dodge(width = 0.6),
  arrow = NULL,
  arrow.fill = NULL,
  lineend = "butt",
  linejoin = "round",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  width = NULL,
  ...
)
}
\description{
geom_segment_JuJu
}
