% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-t_test_barchart.R
\name{geom_text_JuJu}
\alias{geom_text_JuJu}
\title{statistic geom_text}
\usage{
geom_text_JuJu(data, test.type = "t_test")
}
\arguments{
\item{data}{the data}

\item{test.type}{the test type}
}
\description{
statistic geom_text
}
