% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-GeomSegmentJuJu.R
\docType{data}
\name{GeomSegmentJuJu}
\alias{GeomSegmentJuJu}
\title{GeomSegmentJuJu}
\format{
An object of class \code{GeomSegmentJuJu} (inherits from \code{GeomSegment}, \code{Geom}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
GeomSegmentJuJu
}
\description{
GeomSegmentJuJu
}
\keyword{datasets}
