% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-panel.R
\name{image_ggplot_JuJu}
\alias{image_ggplot_JuJu}
\title{image ggplot JuJu}
\usage{
image_ggplot_JuJu(
  magick_image,
  plot.margin = margin(t = 5, b = 5, r = 5, l = 5, unit = "pt"),
  panel.background = element_blank(),
  plot.background = element_blank(),
  aspect.ratio = NULL
)
}
\arguments{
\item{magick_image}{the magick image}
}
\description{
image ggplot JuJu
}
