% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-panel.R
\name{image_read_JuJu}
\alias{image_read_JuJu}
\title{image read as ggplot graph}
\usage{
image_read_JuJu(
  path,
  plot.margin = margin(t = 5, b = 5, r = 5, l = 5, unit = "pt"),
  panel.background = element_blank(),
  plot.background = element_blank()
)
}
\arguments{
\item{path}{the path list with graph name}
}
\description{
image read as ggplot graph
}
