% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-integrated_analysis_pipeline.R
\name{integrated_analysis_pipline_JuJu}
\alias{integrated_analysis_pipline_JuJu}
\title{integrated_analysis_pipeline}
\usage{
integrated_analysis_pipline_JuJu(
  dataset_dt,
  group_col,
  test_col,
  export = FALSE,
  alpha = 0.05
)
}
\arguments{
\item{dataset_dt}{the dataset tabel}

\item{group_col}{the group col}

\item{test_col}{the testing col}
}
\description{
integrated_analysis_pipeline
}
