% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-t_test.R
\name{plot_t_test_barchart_JuJu}
\alias{plot_t_test_barchart_JuJu}
\title{plot t_test bar chart}
\usage{
plot_t_test_barchart_JuJu(t_test_result)
}
\arguments{
\item{t_test_result}{the t_test result datatable}
}
\description{
plot t_test bar chart
}
