% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-t_test_barchart.R
\name{t_test_barchart_JuJu}
\alias{t_test_barchart_JuJu}
\title{t test barchart}
\usage{
t_test_barchart_JuJu(
  t_test_result_JuJu,
  group_order,
  plotting_variable,
  aspect.ratio = sqrt(2),
  legend.position = "top",
  legend.direction = "horizontal",
  scales.free = "free_y",
  plot.facet = TRUE,
  expand.mult = sqrt(2),
  dodge.width = 0.8,
  bar.width = 0.6,
  error.width = 0.2,
  plot.title = "title",
  x_ticks = NULL,
  x_axis_angle = 25,
  x_axis_hjust = 1,
  panel.background = element_blank(),
  plot.background = element_blank(),
  plot.margin = margin(5, 5, 5, 5),
  color_set = c("#6CBAD8", "#96C2D4", "#BAD2E1", "#D8E5F7", "#D8FEFA"),
  x_label = NULL,
  y_label = NULL
)
}
\arguments{
\item{t_test_result_JuJu}{the t test result}

\item{group_order}{the group order}

\item{plotting_variable}{the plotting variable}
}
\description{
t test barchart
}
