% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-general.R
\name{.create_group_pairs}
\alias{.create_group_pairs}
\title{create group pairs}
\usage{
.create_group_pairs(group_list)
}
\arguments{
\item{group_list}{the group list}
}
\description{
create group pairs
}
